requirements.txt
ecdsa
safe-pysha3
requests